libStash/requirements.txt
appdirs==1.4.3
appnope==0.1.2
asgiref==3.3.1
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography
decorator==4.4.2
defusedxml==0.7.0rc2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.5.0
django-crispy-forms==1.9.1
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.21
django-nine==0.2.4
django-phone-field==1.8.1
django-redis==4.12.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
filelock==3.0.12
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
isort==5.5.2
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.3
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2019.3
redis==3.5.3
regex==2020.7.14
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
sendgrid==6.5.0
setuptools-rust==0.11.6
six==1.14.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
stripe==2.55.2
toml==0.10.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.0.9
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
